+++ cognite/pygen/_core/models/fields/primitive.py
@@ -7,7 +7,7 @@ from dataclasses import dataclass
 
 from cognite.client.data_classes import data_modeling as dm
 
-from .base import Field, _to_python_type
+from .base import Field
 
 
 @dataclass(frozen=True)
@@ -134,3 +134,26 @@ class PrimitiveListField(ListFieldCore):
     @property
     def is_list(self) -> bool:
         return True
+
+
+def _to_python_type(type_: dm.DirectRelationReference | dm.PropertyType) -> str:
+    if isinstance(type_, (dm.Int32, dm.Int64)):
+        out_type = "int"
+    elif isinstance(type_, dm.Boolean):
+        out_type = "bool"
+    elif isinstance(type_, (dm.Float32, dm.Float64)):
+        out_type = "float"
+    elif isinstance(type_, dm.Date):
+        out_type = "datetime.date"
+    elif isinstance(type_, dm.Timestamp):
+        out_type = "datetime.datetime"
+    elif isinstance(type_, dm.Json):
+        out_type = "dict"
+    elif isinstance(type_, dm.TimeSeriesReference):
+        out_type = "TimeSeries"
+    elif isinstance(type_, (dm.Text, dm.DirectRelation, dm.CDFExternalIdReference, dm.DirectRelationReference)):
+        out_type = "str"
+    else:
+        raise ValueError(f"Unknown type {type_}")
+
+    return out_type
